backend/pyproject.toml
[project]
name = "todo-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "pydantic",
    "passlib[bcrypt]",
    "python-jose",
    "python-multipart",
    "email-validator",
    "pydantic-settings",
    "PyYAML",
    "python-dotenv",
    "psycopg2-binary"
]
